Refactor: Eliminate floating point and math functions

+++ App/Inc/IPressureSensor.hpp
@@ -11,8 +11,8 @@ public:
     // 初始化气压传感器，加载内部出厂校准寄存器参数，成功返回 Sys::Status::OK
     virtual Sys::Status init() = 0;
     
-    // 获取最新大气压强原始补偿值(Pa) 与 估算海拔高度(m)
-    virtual Sys::Status read(float& pressure, float& altitude) = 0;
+    // 获取最新大气压强原始补偿值(Pa) 与 估算海拔高度(放大10倍, m*10)
+    virtual Sys::Status read(uint32_t& pressure, int32_t& altitude) = 0;
 };
 
 } // namespace App
