+++ App/Inc/ITempHumSensor.hpp
@@ -11,8 +11,8 @@ public:
     // 初始化传感器引脚、总线或内部状态，成功返回 Sys::Status::OK
     virtual Sys::Status init() = 0;
     
-    // 周期获取最新物理数值：温度(℃) 与 相对湿度(%)
-    virtual Sys::Status read(float& temperature, float& humidity) = 0;
+    // 周期获取最新物理数值：温度(放大10倍) 与 相对湿度(放大10倍)
+    virtual Sys::Status read(int32_t& temperature, int32_t& humidity) = 0;
 };
 
 } // namespace App
